demoprac/src/test/java/com/imtiyaaz/tpapppractical/Factories/ClientFactoryTest.java
package com.imtiyaaz.tpapppractical.Factories;

import com.imtiyaaz.tpapppractical.Domain.Client;
import org.junit.Before;
import org.junit.Test;

import java.util.HashMap;
import java.util.Map;

import static org.junit.Assert.assertEquals;

/**
 * Created by Ameer on 2017/08/13.
 */
public class ClientFactoryTest {
    Map<String, String> values;

    @Before
    public void setUp() throws Exception {
        values = new HashMap<String, String>();

        values.put("cName", "John");
        values.put("cEmail", "JohnDoe@gmail.com");
        values.put("cAddress", "34 Some Road SomeCity");
        values.put("cPhone", "0722019397");
        values.put("cCnp", "xxxx");
        values.put("cIdNumber", "88043535435");

    }

    @Test
    public void testGetClient() throws Exception {
        Client client = ClientFactory.getClient(values, 5000);
        System.out.println(client.getcName());
        assertEquals("JimJones", client.getcName());
    }

}